Fixed a bug with variable lenses and removed zero MF matches
I fixed a bug that didn't re-initialize a flag indicating lenses with at least one variable radius, and prevented the use of variable-only stock lens matching.
Some matches give a Merit Function that can't be calculated, and would give an error in the UI. In the ZOS-API the MF value is simply returned as zero. The matches that give a MF value of zero are discarded for this reason. It also means that if a match would legitimately lead to a MF value of zero, it would be ignored, but I think it is unlikely.